fix(bank-user): delete custom SQL migration for roles and permissions init and init data va Java code RolePermissionInitializer class, improve RolePermissionLinkService class add some checks, change all entity ID field Long -> Integer, remove Set<Role> filed from User entity,

+++ services/bank-user/src/main/java/am/banking/system/user/model/repository/PermissionRepository.java
@@ -14,13 +14,13 @@ import reactor.core.publisher.Mono;
  * Time: 00:06:06
  */
 @Repository
-public interface PermissionRepository extends ReactiveCrudRepository<Permission, Long> {
+public interface PermissionRepository extends ReactiveCrudRepository<Permission, Integer> {
     @Query("""
 SELECT p.* FROM user_db.usr.permission p
 JOIN user_db.usr.role_permission rp ON p.id = rp.permission_id
 WHERE rp.role_id = :roleId
 """)
-    Flux<Permission> findAllByRoleId(Long roleId);
+    Flux<Permission> findAllByRoleId(Integer roleId);
 
     @Query("SELECT * FROM user_db.usr.permission WHERE permission_name = :permissionName")
     Mono<Permission> findByPermissionEnum(PermissionEnum permissionName);
